requirements.txt
openai>=1.50.0
python-dotenv>=1.0.1
streamlit>=1.28.0

